first commit
First libnds project; this is a project to both be released eventually to
the public and
a learning attempt. Due to this I will probably not complete it. It is supposed to be a reworking of the Last Night
on Earth board game, though not turn based. Implimented object oriented
design and contains a map, camera class, zombie class, hero class and a
gamecore class(which isn't being used yet). The game as it is makes a
hero, spawns a few zombies and keeps camera centered. Glitches with
sprite palettes.