Main Menu
Added main menu class which the game starts out from. The game now flows
from the main menu to the game and back again. The status of the game is
not reset to default though so it will be like it is on pause. Some other
minor fixes have been implemented.