Score System and saving implemented
While not user side the game has scoring and saving of the scores
implemented. Zombies should not stay blitted on screen anymore either.
Next time user will be able to see scores and fight animations need to be
implemented, will probably update sprites to do this and then game will be
a playable alpha.